Reorganized coordinator files and created a new one for info view
