+++ Modulite/Coordinators/WidgetBuilder/WidgetBuilderCoordinator.swift
@@ -86,7 +86,7 @@ class WidgetBuilderCoordinator: Coordinator {
         router.present(viewController, animated: animated, onDismiss: onDismiss)
     }
     
-    private func presentBackAlertForViewController(
+    func presentBackAlertForViewController(
         _ parentViewController: UIViewController,
         message: String,
         discardAction: @escaping (() -> Void)
@@ -113,110 +113,6 @@ class WidgetBuilderCoordinator: Coordinator {
         alert.addAction(discardChangesAction)
         
         parentViewController.present(alert, animated: true)
-    }
-}
-
-// MARK: - WidgetSetupViewControllerDelegate
-extension WidgetBuilderCoordinator: WidgetSetupViewControllerDelegate {
-    func getPlaceholderName() -> String {
-        .localized(
-            for: .widgetSetupViewMainWidgetNamePlaceholder(number: currentWidgetCount + 1)
-        )
-    }
-    
-    func widgetSetupViewControllerDidPressNext(widgetName: String) {
-        contentBuilder.setWidgetName(widgetName)
-        
-        let viewController = WidgetEditorViewController.instantiate(
-            builder: configurationBuilder,
-            delegate: self
-        )
-        
-        if injectedConfiguration != nil {
-            viewController.loadDataFromBuilder(configurationBuilder)
-            viewController.navigationItem.title = .localized(for: .widgetEditingNavigationTitle)
-            viewController.setIsEditingViewToTrue()
-        }
-        
-        router.present(viewController, animated: true)
-    }
-    
-    func widgetSetupViewControllerDidSelectWidgetStyle(
-        _ controller: WidgetSetupViewController,
-        style: WidgetStyle
-    ) {
-        contentBuilder.setWidgetStyle(style)
-        
-        if let config = injectedConfiguration {
-            config.randomizeWithNewStyle(style)
-        }
-    }
-    
-    func widgetSetupViewControllerDidTapSearchApps(
-        _ parentController: WidgetSetupViewController
-    ) {
-        let router = ModalNavigationRouter(parentViewController: parentController)
-        
-        let coordinator = SelectAppsCoordinator(
-            delegate: self,
-            selectedApps: contentBuilder.getCurrentApps(),
-            router: router
-        )
-        
-        presentChild(coordinator, animated: true) { [weak self] in
-            guard let self = self else { return }
-            
-            parentController.setSetupViewHasAppsSelected(
-                to: !self.contentBuilder.getCurrentApps().isEmpty
-            )
-            
-            parentController.didFinishSelectingApps(
-                apps: self.contentBuilder.getCurrentApps()
-            )
-        }
-    }
-    
-    func widgetSetupViewControllerDidDeselectApp(
-        _ controller: WidgetSetupViewController,
-        app: AppInfo
-    ) {
-        contentBuilder.removeApp(app)
-        
-        if contentBuilder.getCurrentApps().isEmpty {
-            controller.setSetupViewHasAppsSelected(to: false)
-        }
-        
-        guard let config = injectedConfiguration else { return }
-        guard let idx = config.modules.firstIndex(where: { $0.appName == app.name }) else {
-            print("Tried to deselect an app that is not in injected configuration.")
-            return
-        }
-        
-        config.modules.replace(
-            at: idx,
-            with: ModuleConfiguration.empty(
-                style: config.widgetStyle!,
-                at: idx
-            )
-        )
-    }
-    
-    func widgetSetupViewControllerDidPressBack(
-        _ viewController: WidgetSetupViewController,
-        didMakeChanges: Bool
-    ) {
-        if didMakeChanges {
-            presentBackAlertForViewController(
-                viewController,
-                message: .localized(for: .widgetEditingUnsavedChangesAlertMessage)
-            ) { [weak self] in
-                self?.dismiss(animated: true)
-            }
-            
-            return
-        }
-        
-        dismiss(animated: true)
     }
 }
 
@@ -295,5 +191,24 @@ extension WidgetBuilderCoordinator: WidgetEditorViewControllerDelegate {
         ) { [weak self] in
             self?.router.dismissTopViewController(animated: true)
         }
+    }
+    
+    func widgetEditorViewControllerDidPressLayoutInfo(
+        _ viewController: WidgetEditorViewController
+    ) {
+        let router = ModalNavigationRouter(parentViewController: viewController)
+        router.setHasSaveButton(false)
+        
+        let coordinator = TutorialEditWidgetCoordinator(
+            router: router
+        )
+        
+        presentChild(coordinator, animated: true)
+    }
+    
+    func widgetEditorViewControllerDidPressWallpaperInfo(
+        _ viewController: WidgetEditorViewController
+    ) {
+        
     }
 }
